requirements.txt
autopep8==2.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
grpcio==1.51.3
grpcio-tools==1.51.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.22.0
pycodestyle==2.10.0
tomli==2.0.1
Werkzeug==2.2.3
